+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Clients\CommandRunner added
 * Exceptions\E4xx_CommandFailed added
 * Exceptions\E4xx_CommandSucceeded added
+* ProcessRunners\PopenProcessRunner added
 * Requirements\RequireCommandFailed added
 * Requirements\RequireCommandSucceeded added
 * Values\CommandResult added